refactor: Extração da funcionalidade de processamento de perguntas / queries para um arquivo especializado
- E ajuste de configuração para correto funcionamento do Pylint no VS Code
